requirements.txt
Flask==2.2.3
paho-mqtt==1.6.1
PyHaier==0.2.8
pymodbus==2.5.3
pyserial==3.5
requests==2.28.2
schedule==1.2.0
termcolor==2.2.0
waitress==2.1.2
flask-simplelogin==0.1.1
Werkzeug==2.2.3
w1thermsensor==2.0.0
GitPython==3.1.31
